requirements.txt
streamlit
matplotlib
urlextract
seaborn
wordcloud
pandas
emoji
Pillow
plotly